Moved integration tests to new folder, renamed userInfo tests to getUserInfo
